Reorganise auth types ahead of changes
