+++ packages/graphql/src/schema/get-nodes.ts
@@ -24,7 +24,8 @@ import type { Exclude } from "../classes";
 import { Node } from "../classes";
 import type { NodeDirective } from "../classes/NodeDirective";
 import type { QueryOptionsDirective } from "../classes/QueryOptionsDirective";
-import type { Auth, FullText, Neo4jGraphQLCallbacks } from "../types";
+import type { FullText, Neo4jGraphQLCallbacks } from "../types";
+import type { Auth } from "../types/deprecated/auth/auth";
 import getObjFieldMeta from "./get-obj-field-meta";
 import parsePluralDirective from "./parse/parse-plural-directive";
 import { parseQueryOptionsDirective } from "./parse/parse-query-options-directive";
