+++ packages/graphql/src/schema/resolvers/wrapper-utils.ts
@@ -19,7 +19,8 @@
 
 import { Neo4jGraphQLAuthenticationError } from "../../classes";
 import type { SubscriptionContext } from "./subscriptions/types";
-import type { Context, JwtPayload, Neo4jGraphQLAuthPlugin } from "../../types";
+import type { Context, Neo4jGraphQLAuthPlugin } from "../../types";
+import type { JwtPayload } from "../../types/deprecated/auth/jwt-payload";
 
 export async function decodeToken(
     token: string | undefined,
