+++ packages/graphql/src/utils/execute.test.ts
@@ -19,7 +19,7 @@
 
 import type { Driver } from "neo4j-driver";
 import type { Neo4jGraphQL } from "../classes";
-import type { AuthContext } from "../types";
+import type { AuthContext } from "../types/deprecated/auth/auth-context";
 import {
     CypherConnectComponentsPlanner,
     CypherExpressionEngine,
