+++ packages/graphql/tests/utils/builders/context-builder.ts
@@ -22,7 +22,8 @@ import type { ResolveTree } from "graphql-parse-resolve-info";
 import { GraphQLSchema } from "graphql";
 import { Neo4jGraphQL } from "../../../src/classes";
 import type { Neo4jDatabaseInfo } from "../../../src/classes/Neo4jDatabaseInfo";
-import type { AuthContext, Context } from "../../../src/types";
+import type { Context } from "../../../src/types";
+import type { AuthContext } from "../../../src/types/deprecated/auth/auth-context";
 import { Builder } from "./builder";
 import { Executor } from "../../../src/classes/Executor";
 import { Neo4jGraphQLSchemaModel } from "../../../src/schema-model/Neo4jGraphQLSchemaModel";
